requirements.txt
opencv-python
pytesseract
sentence-transformers
faiss-cpu
openai
watchdog
numpy
python-dotenv
huggingface_hub
pillow
pillow-heif